+++ Examples/Resources/SqlServerLogin/2-RemoveLogin.ps1
@@ -16,29 +16,29 @@ Configuration Example
     node localhost {
         SqlServerLogin Remove_WindowsUser
         {
-            Ensure          = 'Absent'
-            Name            = 'CONTOSO\WindowsUser'
-            LoginType       = 'WindowsUser'
-            SQLServer       = 'SQLServer'
-            SQLInstanceName = 'DSC'
+            Ensure       = 'Absent'
+            Name         = 'CONTOSO\WindowsUser'
+            LoginType    = 'WindowsUser'
+            ServerName   = 'TestServer.company.local'
+            InstanceName = 'DSC'
         }
 
         SqlServerLogin Remove_WindowsGroup
         {
-            Ensure          = 'Absent'
-            Name            = 'CONTOSO\WindowsGroup'
-            LoginType       = 'WindowsGroup'
-            SQLServer       = 'SQLServer'
-            SQLInstanceName = 'DSC'
+            Ensure       = 'Absent'
+            Name         = 'CONTOSO\WindowsGroup'
+            LoginType    = 'WindowsGroup'
+            ServerName   = 'TestServer.company.local'
+            InstanceName = 'DSC'
         }
 
         SqlServerLogin Remove_SqlLogin
         {
-            Ensure          = 'Absent'
-            Name            = 'SqlLogin'
-            LoginType       = 'SqlLogin'
-            SQLServer       = 'SQLServer'
-            SQLInstanceName = 'DSC'
+            Ensure       = 'Absent'
+            Name         = 'SqlLogin'
+            LoginType    = 'SqlLogin'
+            ServerName   = 'TestServer.company.local'
+            InstanceName = 'DSC'
         }
     }
 }